requirements.txt
pyyaml
openai
python-dotenv
requests
beautifulsoup4
fastapi
uvicorn
websockets
pydantic

# Testing requirements
pytest
pytest-asyncio
pytest-cov
pytest-mock
psutil